First draft of multistep interaction create
