+++ src/feedback/CreateDetails.vue
@@ -54,7 +54,7 @@ onMounted(async () => {
     !feedbackSession.isSingle &&
     !feedbackSession.useTemplate
   ) {
-    router.push("/feedback/create-type");
+    router.push("/feedback/create/type");
   }
 });
 </script>
