+++ src/feedback/CreateOrganisers.vue
@@ -155,7 +155,7 @@ onMounted(async () => {
     !feedbackSession.isSingle &&
     !feedbackSession.useTemplate
   ) {
-    router.push("/feedback/create-type");
+    router.push("/feedback/create/type");
   }
 });
 </script>
