+++ src/feedback/CreateType.vue
@@ -140,7 +140,7 @@ const selectTemplate = async (forceOn) => {
   nextDisabled.value = false;
 };
 
-const loadTemplateDetails = async () => {
+const fetchTemplate = async () => {
   try {
     const response = await api("feedback/loadUpdateSession", {
       id: feedbackSession.id,
@@ -242,7 +242,7 @@ const loadTemplate = async () => {
   });
 
   if (isConfirmed) {
-    if (await loadTemplateDetails()) return true;
+    if (await fetchTemplate()) return true;
   }
   return false;
 };
@@ -282,269 +282,263 @@ onMounted(async () => {
 
 <template>
   <div class="container d-flex flex-column align-items-center">
-    <div class="options-container w-100">
-      <h1 class="text-center display-4">Feedback</h1>
-      <div class="text-center">
-        <p>Let's get started creating your feedback request</p>
-      </div>
-      <!--option buttons-->
-      <div class="d-flex flex-column align-items-stretch">
-        <!--single-->
-        <div
-          class="border border-2 rounded p-1"
-          :class="{
-            'border-teal':
-              !feedbackSession.isSeries && !feedbackSession.useTemplate,
-          }"
-          @click="selectSingle(false)"
-        >
-          <div class="d-flex justify-content-between align-items-center">
-            <span>Collect feedback on a teaching event with one session</span>
-            <button
-              class="btn"
-              :class="{
-                'btn-teal': feedbackSession.isSingle,
-              }"
-            >
-              <font-awesome-icon
-                :icon="['far', 'square-check']"
-                v-if="feedbackSession.isSingle"
-              />
-              <font-awesome-icon :icon="['far', 'square']" v-else />
-            </button>
-          </div>
-          <!--single help-->
-          <div class="collapse form-text mx-1" id="singleHelp">
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />Create a simple feedback form for a single session.</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />Certificate of attendance option.</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />Register of attendance option.</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />Add additional custom questions if required.</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />Add as many co-organisers as you need.</span
-            ><br />
-          </div>
+    <h1 class="text-center display-4">Feedback</h1>
+    <div class="text-center">
+      <p>Let's get started creating your feedback request</p>
+    </div>
+    <!--option buttons-->
+    <div class="d-flex flex-column align-items-stretch w-100">
+      <!--single-->
+      <div
+        class="border border-2 rounded p-1"
+        :class="{
+          'border-teal':
+            !feedbackSession.isSeries && !feedbackSession.useTemplate,
+        }"
+        @click="selectSingle(false)"
+      >
+        <div class="d-flex justify-content-between align-items-center">
+          <span>Collect feedback on a teaching event with one session</span>
+          <button
+            class="btn"
+            :class="{
+              'btn-teal': feedbackSession.isSingle,
+            }"
+          >
+            <font-awesome-icon
+              :icon="['far', 'square-check']"
+              v-if="feedbackSession.isSingle"
+            />
+            <font-awesome-icon :icon="['far', 'square']" v-else />
+          </button>
         </div>
-
-        <!--series-->
-        <div
-          class="border border-2 rounded p-1 mt-4"
-          :class="{
-            'border-teal':
-              !feedbackSession.isSingle && !feedbackSession.useTemplate,
-          }"
-          @click="selectSeries(false)"
-        >
-          <div class="d-flex justify-content-between align-items-center">
-            <span
-              >Collect feedback on a teaching event with multiple sessions</span
-            >
-            <button
-              class="btn"
-              :class="{
-                'btn-teal': feedbackSession.isSeries,
-              }"
-            >
-              <font-awesome-icon
-                :icon="['far', 'square-check']"
-                v-if="feedbackSession.isSeries"
-              />
-              <font-awesome-icon :icon="['far', 'square']" v-else />
-            </button>
-          </div>
-          <!--series help-->
-          <div class="collapse form-text mx-1" id="seriesHelp">
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />
-              Gather feedback on multiple sessions using a single form.</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />
-              Attendees wil be asked to provide feedback for each session you
-              add and for the event as a whole.</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />
-              You can give facilitators of each session access to view the
-              feedback (just for their session).</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />Certificate of attendance option.</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />Register of attendance option.</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />Add additional custom questions if required (just for the
-              overall feedback, not each session).</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />Add as many co-organisers as you need.</span
-            ><br />
-          </div>
-        </div>
-
-        <!--template-->
-        <div
-          class="border border-2 rounded p-1 mt-4"
-          :class="{
-            'border-teal':
-              !feedbackSession.isSeries && !feedbackSession.isSingle,
-          }"
-          @click="selectTemplate(false)"
-        >
-          <div class="d-flex justify-content-between align-items-center">
-            <span>Use a previous event as a template to collect feedback</span>
-            <button
-              class="btn"
-              :class="{
-                'btn-teal': feedbackSession.useTemplate,
-              }"
-            >
-              <font-awesome-icon
-                :icon="['far', 'square-check']"
-                v-if="feedbackSession.useTemplate"
-              />
-              <font-awesome-icon :icon="['far', 'square']" v-else />
-            </button>
-          </div>
-          <!--template help-->
-          <div class="collapse form-text mx-1" id="templateHelp">
-            <span
-              ><font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />Use a previous feedback form as a template to save time.</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />You will need the session ID and PIN.</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />You can make changes before creating the new form.</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />The form used as a template will not be changed.</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />Certificate of attendance option.</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />Register of attendance option.</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />Add additional custom questions if required.</span
-            ><br />
-            <span>
-              <font-awesome-icon
-                :icon="['fas', 'check']"
-                style="color: #17a2b8"
-                class="me-1"
-              />Add as many co-organisers as you need.</span
-            ><br />
-          </div>
+        <!--single help-->
+        <div class="collapse form-text mx-1" id="singleHelp">
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />Create a simple feedback form for a single session.</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />Certificate of attendance option.</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />Register of attendance option.</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />Add additional custom questions if required.</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />Add as many co-organisers as you need.</span
+          ><br />
         </div>
       </div>
 
-      <!--back/next button-->
-      <div class="d-flex justify-content-evenly my-3">
-        <button
-          class="btn btn-teal btn-lg me-2 mb-2"
-          id="next"
-          @click="next"
-          :disabled="nextDisabled"
-        >
-          Continue
-        </button>
+      <!--series-->
+      <div
+        class="border border-2 rounded p-1 mt-4"
+        :class="{
+          'border-teal':
+            !feedbackSession.isSingle && !feedbackSession.useTemplate,
+        }"
+        @click="selectSeries(false)"
+      >
+        <div class="d-flex justify-content-between align-items-center">
+          <span
+            >Collect feedback on a teaching event with multiple sessions</span
+          >
+          <button
+            class="btn"
+            :class="{
+              'btn-teal': feedbackSession.isSeries,
+            }"
+          >
+            <font-awesome-icon
+              :icon="['far', 'square-check']"
+              v-if="feedbackSession.isSeries"
+            />
+            <font-awesome-icon :icon="['far', 'square']" v-else />
+          </button>
+        </div>
+        <!--series help-->
+        <div class="collapse form-text mx-1" id="seriesHelp">
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />
+            Gather feedback on multiple sessions using a single form.</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />
+            Attendees wil be asked to provide feedback for each session you add
+            and for the event as a whole.</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />
+            You can give facilitators of each session access to view the
+            feedback (just for their session).</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />Certificate of attendance option.</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />Register of attendance option.</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />Add additional custom questions if required (just for the overall
+            feedback, not each session).</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />Add as many co-organisers as you need.</span
+          ><br />
+        </div>
       </div>
+
+      <!--template-->
+      <div
+        class="border border-2 rounded p-1 mt-4"
+        :class="{
+          'border-teal': !feedbackSession.isSeries && !feedbackSession.isSingle,
+        }"
+        @click="selectTemplate(false)"
+      >
+        <div class="d-flex justify-content-between align-items-center">
+          <span>Use a previous event as a template to collect feedback</span>
+          <button
+            class="btn"
+            :class="{
+              'btn-teal': feedbackSession.useTemplate,
+            }"
+          >
+            <font-awesome-icon
+              :icon="['far', 'square-check']"
+              v-if="feedbackSession.useTemplate"
+            />
+            <font-awesome-icon :icon="['far', 'square']" v-else />
+          </button>
+        </div>
+        <!--template help-->
+        <div class="collapse form-text mx-1" id="templateHelp">
+          <span
+            ><font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />Use a previous feedback form as a template to save time.</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />You will need the session ID and PIN.</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />You can make changes before creating the new form.</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />The form used as a template will not be changed.</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />Certificate of attendance option.</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />Register of attendance option.</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />Add additional custom questions if required.</span
+          ><br />
+          <span>
+            <font-awesome-icon
+              :icon="['fas', 'check']"
+              style="color: #17a2b8"
+              class="me-1"
+            />Add as many co-organisers as you need.</span
+          ><br />
+        </div>
+      </div>
+    </div>
+
+    <!--back/next button-->
+    <div class="d-flex justify-content-evenly my-3">
+      <button
+        class="btn btn-teal btn-lg me-2 mb-2"
+        id="next"
+        @click="next"
+        :disabled="nextDisabled"
+      >
+        Continue
+      </button>
     </div>
   </div>
 </template>
 
 <style>
 .container {
-  max-width: 750px;
-}
-.options-container {
   max-width: 750px;
 }
 </style>
