+++ todo
@@ -1,5 +1,4 @@
 Current:
-Check how edit works with new multistep form
 
 To do
 Preview certificate
